alert-service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
kafka-python